# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MindPal API
  version: 5.0.0
  description: >
    Source of truth for HTTP. FastAPI may only expose operations listed here.
    Brain (Obsidian workspace HTTP) is retired; memory graph is the durable store.
...
paths:
...
  /api/user/profile:
    get:
      operationId: identityGetProfile
      tags: [identity]
      x-mindpal:
        owner: identity
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: false
      responses:
        "200": { description: Profile }
        "401": { description: Sign in required }
    patch:
      operationId: identityPatchProfile
      tags: [identity]
      x-mindpal:
        owner: identity
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: true
      responses:
        "200": { description: Profile }
        "401": { description: Sign in required }
  /api/user/insights:
    get:
      operationId: identityGetInsights
      tags: [identity]
      x-mindpal:
        owner: identity
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: false
      responses:
        "200": { description: Insights }
        "401": { description: Sign in required }
  /api/user/wellness-timeline:
    get:
      operationId: identityGetWellnessTimeline
      tags: [identity]
      summary: Mood, themes, and life events from saved memory and synced chats
      description: >
        Authenticated reflection of the signed-in account's saved memory atoms and
        synced user turns. Coarse valence labels (heavier / mixed / lighter) are
        derived from the person's words. This is not PHQ-9, GAD-7, or a clinical
        score. Delete `/api/user/data` removes the source data. Guests receive 401.
      x-mindpal:
        owner: identity
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: false
      responses:
        "200": { description: Wellness timeline derived from stored memory and chats }
        "401": { description: Sign-in required }
  /api/user/export:
    get:
      operationId: identityExport
      tags: [identity]
      summary: Download signed-in account data
      description: >
        Authenticated JSON of server-stored profile, memory graph, and synced
        chat sessions. Does not include this browser's local history or guest facts.
      x-mindpal:
        owner: identity
        status: preview
        audience: user
        ai-access: hidden
        client: web
        sunset: null
        side-effects: false
      responses:
        "200": { description: JSON export of profile, memory, and synced chats }
        "401": { description: Sign-in required }
  /api/user/data:
    delete:
      operationId: identityDeleteData
      tags: [identity]
      summary: Delete signed-in account data
      description: >
        Permanently deletes the server profile, memory graph, and synced chats
        for the signed-in account. Local browser chat history is not removed.
      x-mindpal:
        owner: identity
        status: preview
        audience: user
        ai-access: hidden
        client: web
        sunset: null
        side-effects: true
      responses:
        "200": { description: Deleted server profile, memory, and synced chats }
        "401": { description: Sign-in required }
  /api/memory/graph:
    get:
      operationId: memoryGetGraph
      tags: [memory]
      summary: Canonical durable memory graph
      x-mindpal:
        owner: memory
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: false
      responses:
        "200": { description: Graph }
    put:
      operationId: memoryPutGraph
      tags: [memory]
      x-mindpal:
        owner: memory
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: true
      responses:
        "200": { description: Graph }
        "401": { description: Sign in required }
  /api/memory/graph/items/{atom_id}:
    patch:
      operationId: memoryPatchGraphItem
      tags: [memory]
      summary: Update stored text for one memory atom
      x-mindpal:
        owner: memory
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: true
      responses:
        "200": { description: Graph after update }
        "401": { description: Sign-in required }
        "404": { description: Atom not saved }
        "422": { description: Empty memory text }
    delete:
      operationId: memoryDeleteGraphItem
      tags: [memory]
      x-mindpal:
        owner: memory
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: true
      responses:
        "200": { description: Graph after delete }
        "401": { description: Sign in required }
...
  /api/memory/summary/refresh:
    post:
      operationId: memoryRefreshSummary
      tags: [memory]
      summary: >-
        Recompute the stored memory summary from saved atoms. The body is
        optional; with no body the summary is rebuilt from what is saved, and
        with {"summary": "..."} that text is stored instead. Deterministic, not
        a model call.
      x-mindpal:
        owner: memory
        status: preview
        audience: user
        ai-access: schema
        client: web
        sunset: null
        side-effects: true
      responses:
        "200": { description: Summary }
        "401": { description: Sign in required }
...